Revert unnecessary change

+++ sdk/cosmos/azure-cosmos/tests/test_vector_policy.py
@@ -462,7 +462,7 @@ class TestVectorPolicy(unittest.TestCase):
             "vectorEmbeddings": [
                 {
                     "path": "/vector1",
-                    "dataType": "float16",
+                    "dataType": "float32",
                     "dimensions": 256,
                     "distanceFunction": "euclidean"
                 }]}
